little refactoring, drop more obsolete HardTypes
#21
