+++ type-registry/src/main/scala/dev/rudiments/types/registry/Application.scala
@@ -7,7 +7,7 @@ import akka.stream.ActorMaterializer
 import com.typesafe.config.ConfigFactory
 import com.typesafe.scalalogging.LazyLogging
 import dev.rudiments.hardcore.data.CRUD.Create
-import dev.rudiments.hardcore.data.{DataHttpPort, DataMemoryAdapter}
+import dev.rudiments.hardcore.data.DataMemoryAdapter
 import dev.rudiments.hardcore.http.RootRouter
 import dev.rudiments.hardcore.types._
 import dev.rudiments.types.registry.module.TypeHttpPort
@@ -24,13 +24,11 @@ object Application extends App with LazyLogging {
   implicit val ec: ExecutionContext = actorSystem.dispatcher
   implicit val mat: ActorMaterializer = ActorMaterializer()
 
-  implicit val t: HardType[Type] = HardType[Type]
-
   try {
     val config = ConfigFactory.load()
     val db = new DataMemoryAdapter[Type]
-    db(Create(ID("Example"), HardType[Example]))
-    db(Create(ID("Sample"), HardType[Sample]))
+    db(Create(ID("CompatibleDTO"), HardType[CompatibleDTO]))
+    db(Create(ID("CompatiblePlainDTO"), HardType[CompatiblePlainDTO]))
 
     new RootRouter(config, TypeHttpPort("types", db)).bind()
   } catch {
@@ -44,56 +42,4 @@ object Application extends App with LazyLogging {
       }
   }
 
-  sealed trait MyEnum extends EnumEntry
-  object MyEnum extends Enum[MyEnum] {
-    override def values: immutable.IndexedSeq[MyEnum] = findValues
-
-    case object One extends MyEnum
-    case object Two extends MyEnum
-    case object Red extends MyEnum
-  }
-
-  case class Example(
-    id: Long,
-    name: String,
-    comment: Option[String] = None,
-    n: Int = Int.MaxValue,
-    array: Seq[Int],
-    arrayWithDefault: Seq[Int] = Seq.empty,
-    question: List[Int] = List(42),
-    when: Timestamp = Defaults.now,
-    date: Option[Date] = Some(Defaults.today),
-    mapIntDate: Map[Int, Date] = Map.empty,
-    sample: Sample,
-    optSample: Option[Sample],
-    seqSample: Seq[Sample] = Seq.empty,
-    setSample: Set[Sample] = Set.empty,
-    mapSample: Map[String, Sample] = Map.empty,
-    multimapSample: Map[String, Set[Sample]] = Map.empty,
-    deepMapSample: Map[String, Map[String, Set[Sample]]] = Map.empty,
-  ) extends DTO
-
-  case class Sample(
-    string: String,
-    optString: Option[String],
-    defaultString: String = "default",
-    defaultOptString: Option[String] = None,
-    listOfStrings: Seq[String],
-    int: Int,
-    optInt: Option[Int],
-    double: Double,
-    optDouble: Option[Double],
-    long: Long,
-    optLong: Option[Long],
-    decimal: BigDecimal,
-    optDecimal: Option[BigDecimal],
-    timestamp: Timestamp,
-    optTimestamp: Option[Timestamp],
-    date: Date,
-    optDate: Option[Date],
-    time: Time,
-    optTime: Option[Time],
-    enum: MyEnum,
-    optEnum: Option[MyEnum]
-  ) extends DTO
 }
