+++ type-registry/src/main/scala/dev/rudiments/types/registry/module/TypeHttpPort.scala
@@ -1,10 +1,9 @@
 package dev.rudiments.types.registry.module
 
 import dev.rudiments.hardcore.data.{DataHttpPort, DataSkill}
-import dev.rudiments.hardcore.types.{HardType, ID, Type}
+import dev.rudiments.hardcore.types.{ID, Type}
 
 object TypeHttpPort {
-  private implicit val t: HardType[Type] = HardType[Type]
 
   import dev.rudiments.hardcore.http.CirceSupport._
   def apply(prefix: String, f: DataSkill[Type]): DataHttpPort[Type, String] = new DataHttpPort[Type, String](
